requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database - MySQL
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
alembic==1.12.1

# PDF Processing
pdfplumber==0.10.3
PyMuPDF==1.23.8
pytesseract==0.3.10
pdf2image==1.16.3

# ML/NLP
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl

# Task Queue (optional for production)
celery==5.3.4
redis==5.0.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# API & Utilities
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
httpx==0.25.2